detection-function/src/requirements.txt
boto3==1.11.11
pillow==6.2.2
